+++ exgen/src/main/java/exercisegenerator/structures/graphs/petrinets/PetriTransition.java
@@ -2,4 +2,4 @@ package exercisegenerator.structures.graphs.petrinets;
 
 import java.util.*;
 
-public record PetriTransition(String label, int x, int y, Map<Integer, Integer> from, Map<Integer, Integer> to) {}
+public record PetriTransition(String label, double x, double y, Map<Integer, Integer> from, Map<Integer, Integer> to) {}
